reordered components folder, added comment regarding moving api call to BE + prettier
