+++ src/components/GameLayout.tsx
